refactor: move logic for campaign pages handling to ManualPageImporter and PreviewPageImporter classes
